Version synchronization
All app*.py versions brought together.
